% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer.R
\name{answer}
\alias{answer}
\title{Stores variable calls in a data table of answers}
\usage{
answer(..., label = "")
}
\arguments{
\item{...}{Arbitrary sequence of variables or function calls to be stored}

\item{label}{String to label values of answers for later checking}
}
\value{
A data table of answers to be checked against solutions
}
\description{
This function stores an arbitrary number of variables
and function calls to be stored in a data table for check()
}
