% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Checks data table of answers with table of solutions}
\usage{
check()
}
\value{
Data table with results rendered
}
\description{
This function compares the values in the answer column of
the solutions_table with those in the check_table. It returns
a table with booleans indicating matches between answer provided
and solution set.
}
