% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_check.R
\name{initialize_check}
\alias{initialize_check}
\title{Initializes data table for student answers}
\usage{
initialize_check()
}
\value{
Blank data table check_table
}
\description{
This function creates a blank data table for
answers to be filled into.
}
