% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_solution.R
\name{initialize_solution}
\alias{initialize_solution}
\title{Initializes data tables}
\usage{
initialize_solution(solutions_path = "", url = FALSE)
}
\value{
Blank data table
}
\description{
This function creates a blank data table for
answers to be filled into.
}
