% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_solutions.R
\name{publish_solutions}
\alias{publish_solutions}
\title{Writes table of solutions to storage}
\usage{
publish_solutions(path = getwd(), file = "solution.RData")
}
\arguments{
\item{path}{Export file path}

\item{file}{Export file name}
}
\value{
.Rdata of solutions table
}
\description{
This function saves a data table of solutions
to be stored online and accessed by students privately
}
